services/engine/requirements.txt
aiohttp==3.7.2
alembic==1.4.3
amqp==2.6.1
APScheduler==3.6.3
async-timeout==3.0.1
attrs==20.3.0
beautifulsoup4==4.9.3
billiard==3.6.3.0
celery==4.3.0
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
configobj==5.0.6
Cython==0.29.21
Flask==1.1.2
Flask-Migrate==2.4.0
Flask-Redis==0.3.0
Flask-SQLAlchemy==2.3.2
gevent==1.4.0
greenlet==0.4.15
grpcio==1.33.2
grpcio-tools==1.33.2
gunicorn==19.9.0
html2text==2020.1.16
idna==2.8
importlib-metadata==2.0.0
itsdangerous==1.1.0
Jinja2==2.11.2
kombu==4.6.11
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==2.19.2
multidict==5.0.2
Pillow==8.2.0
plumbum==1.6.9
protobuf==3.12.2 --no-binary protobuf
psycopg2-binary==2.7.6.1
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.4
raven==6.10.0
redis==3.5.3
requests==2.22.0
rpyc==4.1.5
six==1.15.0
soupsieve==2.1
SQLAlchemy==1.3.20
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.25.11
vine==1.3.0
webargs==4.0.0
Werkzeug==1.0.1
yarl==1.6.3
zipp==3.4.0
